Set fallback view offset for monsters
Some monster models don't have the eye position set. This messes up with monster's ability to see enemy and check for attacks.
Example: Big Momma model doesn't have the eye position, so it doesn't "see" the enemy when he's on a different height and thus doesn't perform the spit attack even if it's possible.