Let monsters attack and knock off enemies that stand on top of their head.
This is mostly to prevent player abusing standing on the monster's head.
We also rewrote how the distance to the enemy is calculated to let the melee attack checks pass when the enemy stands on the top of the attacker.
Otherwise some checks fail as the check distance is often shorter than the monster's height.