Fix monsters looping in the establish line of fire schedule without movement
If enemy LKP happened to be too far from the enemy position, the monster's route to the enemy
gets refreshed with the route to the LKP.
If LKP is too close to the monster's current position the monster will complete the schedule
without actually moving. Then AI might select the same schedule again and repeat.
We're fixing it by checking whether the monster is already close to the enemy LKP.
If it's the case we reuse the first point of previously built route as enemy LKP.